Fixed readwrite.py type hints

+++ pg/datatypes.py
@@ -11,8 +11,8 @@ from typing import NamedTuple, TypedDict
 
 
 class Coordinate(NamedTuple):
-    x: int
-    y: int
+    x: float
+    y: float
 
     def __repr__(self) -> str:
         return f"({self.x}, {self.y})"
@@ -62,3 +62,5 @@ class District(TypedDict):
 
 
 # DON'T LIMIT WHAT GETS EXPORTED.
+
+### END ###
